Moving and adapting ownership to helpers to be able to implement in different scenarios
